+++ src/ezdxf/math/__init__.py
@@ -1,10 +1,10 @@
 # Purpose: math and construction tools
 # Created: 27.03.2010, 2018 integrated into ezdxf
-# Copyright (c) 2010-2019 Manfred Moitzi
+# Copyright (c) 2010-2020, Manfred Moitzi
 # License: MIT License
 from .construct2d import (
     is_close_points, closest_point, convex_hull, intersection_line_line, distance_point_line,
-    is_point_on_line, is_point_in_polygon,
+    is_point_on_line, is_point_in_polygon, is_point_left_of_line, point_to_line_relation,
 )
 from .vector import Vector, Vec2, X_AXIS, Y_AXIS, Z_AXIS, NULLVEC
 from .matrix44 import Matrix44